+++ l2geth/core/vm/errors.go
@@ -27,4 +27,5 @@ var (
 	ErrInsufficientBalance      = errors.New("insufficient balance for transfer")
 	ErrContractAddressCollision = errors.New("contract address collision")
 	ErrNoCompatibleInterpreter  = errors.New("no compatible interpreter")
+	ErrGasUintOverflow          = errors.New("gas uint64 overflow")
 )