+++ packages/batch-submitter/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.4.15
+
+### Patch Changes
+
+- ae4a90d9: Adds a fix for the BSS to account for new timestamp logic in L2Geth
+- ca547c4e: Import performance to not couple batch submitter to version of nodejs that has performance as a builtin
+- Updated dependencies [ad94b9d1]
+  - @eth-optimism/core-utils@0.7.5
+  - @eth-optimism/contracts@0.5.10
+
 ## 0.4.14
 
 ### Patch Changes