+++ packages/core-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @eth-optimism/core-utils
 
+## 0.7.5
+
+### Patch Changes
+
+- ad94b9d1: test/docs: Improve docstrings and tests for utils inside of hex-strings.ts
+
 ## 0.7.4
 
 ### Patch Changes